refactor: generalize yaml update function
